+++ include/beast/http.hpp
@@ -12,6 +12,7 @@
 
 #include <beast/http/basic_parser.hpp>
 #include <beast/http/buffer_body.hpp>
+#include <beast/http/chunk_encode.hpp>
 #include <beast/http/dynamic_body.hpp>
 #include <beast/http/empty_body.hpp>
 #include <beast/http/error.hpp>