+++ include/beast/http/detail/chunk_encode.hpp
@@ -8,6 +8,8 @@
 #ifndef BEAST_HTTP_DETAIL_CHUNK_ENCODE_HPP
 #define BEAST_HTTP_DETAIL_CHUNK_ENCODE_HPP
 
+#include <beast/core/type_traits.hpp>
+#include <beast/http/type_traits.hpp>
 #include <boost/asio/buffer.hpp>
 #include <algorithm>
 #include <array>
@@ -17,15 +19,56 @@ namespace beast {
 namespace http {
 namespace detail {
 
+struct chunk_extensions
+{
+    virtual ~chunk_extensions() = default;
+    virtual boost::asio::const_buffers_1 str() = 0;
+};
+    
+template<class ChunkExtensions>
+struct chunk_extensions_impl : chunk_extensions
+{
+    ChunkExtensions ext_;
+
+    chunk_extensions_impl(ChunkExtensions&& ext)
+        : ext_(std::move(ext))
+    {
+    }
+
+    chunk_extensions_impl(ChunkExtensions const& ext)
+        : ext_(ext)
+    {
+    }
+
+    boost::asio::const_buffers_1
+    str() override
+    {
+        auto const s = ext_.str();
+        return {s.data(), s.size()};
+    }
+};
+
+template<class T, class = void>
+struct is_chunk_extensions : std::false_type {};
+
+template<class T>
+struct is_chunk_extensions<T, beast::detail::void_t<decltype(
+    std::declval<string_view&>() = std::declval<T&>().str(),
+        (void)0)>> : std::true_type
+{
+};
+
+//------------------------------------------------------------------------------
+
 /** A buffer sequence containing a chunk-encoding header
 */
-class chunk_header
+class chunk_size
 {
 public:
     // Storage for the longest hex string we might need
     class value_type
     {
-        friend class chunk_header;
+        friend class chunk_size;
 
         // First byte holds the length
         char buf_[1 + 2 * sizeof(std::size_t)];
@@ -63,14 +106,13 @@ public:
 
     using const_iterator = value_type const*;
 
-    chunk_header(chunk_header const& other) = default;
+    chunk_size(chunk_size const& other) = default;
 
     /** Construct a chunk header
 
         @param n The number of octets in this chunk.
     */
-    explicit
-    chunk_header(std::size_t n)
+    chunk_size(std::size_t n)
     {
         value_.prepare(n);
     }
@@ -93,7 +135,7 @@ private:
 
 template<class>
 void
-chunk_header::
+chunk_size::
 value_type::
 prepare(std::size_t n)
 {
@@ -101,6 +143,8 @@ prepare(std::size_t n)
     auto it = to_hex(last, n);
     buf_[0] = static_cast<char>(last - it);
 }
+
+//------------------------------------------------------------------------------
 
 /// Returns a buffer sequence holding a CRLF for chunk encoding
 inline
@@ -113,10 +157,99 @@ chunk_crlf()
 /// Returns a buffer sequence holding a final chunk header
 inline
 boost::asio::const_buffers_1
-chunk_final()
+chunk_last()
 {
     return {"0\r\n", 3};
 }
+
+//------------------------------------------------------------------------------
+
+template<class = void>
+struct chunk_crlf_iter_type
+{
+    class value_type
+    {
+        char const s[2] = {'\r', '\n'};
+        
+    public:
+        value_type() = default;
+
+        operator
+        boost::asio::const_buffer() const
+        {
+            return {s, sizeof(s)};
+        }
+    };
+    static value_type value;
+};
+
+template<class T>
+typename chunk_crlf_iter_type<T>::value_type
+chunk_crlf_iter_type<T>::value;
+
+using chunk_crlf_iter = chunk_crlf_iter_type<void>;
+
+//------------------------------------------------------------------------------
+
+template<class = void>
+struct chunk_size0_iter_type
+{
+    class value_type
+    {
+        char const s[3] = {'0', '\r', '\n'};
+        
+    public:
+        value_type() = default;
+
+        operator
+        boost::asio::const_buffer() const
+        {
+            return {s, sizeof(s)};
+        }
+    };
+    static value_type value;
+};
+
+template<class T>
+typename chunk_size0_iter_type<T>::value_type
+chunk_size0_iter_type<T>::value;
+
+using chunk_size0_iter = chunk_size0_iter_type<void>;
+
+struct chunk_size0
+{
+    using value_type = chunk_size0_iter::value_type;
+
+    using const_iterator = value_type const*;
+
+    const_iterator
+    begin() const
+    {
+        return &chunk_size0_iter::value;
+    }
+
+    const_iterator
+    end() const
+    {
+        return begin() + 1;
+    }
+};
+
+//------------------------------------------------------------------------------
+
+template<class T,
+    bool = is_fields<T>::value>
+struct buffers_or_fields
+{
+    using type = typename
+        T::reader::const_buffers_type;
+};
+
+template<class T>
+struct buffers_or_fields<T, false>
+{
+    using type = T;
+};
 
 } // detail
 } // http